backend/requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
boto3==1.34.0
strands==0.1.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
